pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentprobe"
version = "0.1.0"
description = "Open-source red-teaming framework for AI agent systems"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "AgentProbe Contributors" }]
keywords = ["security", "red-team", "ai-agents", "llm", "penetration-testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "ag2>=0.4",
    "langgraph>=0.2",
    "langchain-anthropic>=0.3",
    "langchain-groq>=0.2",
    "anthropic>=0.40",
    "groq>=0.9",
    "pydantic>=2.0",
    "scikit-learn>=1.4",
    "sentence-transformers>=3.0",
    "streamlit>=1.38",
    "plotly>=5.0",
    "jinja2>=3.1",
    "click>=8.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "joblib>=1.3",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[project.scripts]
agentprobe = "agentprobe.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentprobe*"]
